+++ app/client/intro/intro_modal.js
@@ -1,3 +1,4 @@
+import {request_notifications} from '../requests/request_notifications.js'
 
 export function intro_modal()
 {
@@ -56,6 +57,9 @@ export function intro_modal()
     open_link_in_chrome('link-panel-tutorials')
     open_link_in_chrome('link-panel-downloads')
     open_link_in_chrome('link-panel-stripe-portal')
+
+    // Load the notifications
+    request_notifications()
 }